refactor: 일부 클래스들 main 패키지로 이동
