+++ src/main/kotlin/gamemap/CellState.kt
@@ -1,0 +1,5 @@
+package gamemap
+
+enum class CellState {
+    Open, Close
+}
