+++ src/test/kotlin/gamemap/CellTest.kt
@@ -1,3 +1,5 @@
+import gamemap.Cell
+import gamemap.CellState
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
@@ -5,7 +7,7 @@ import io.kotest.matchers.string.shouldContain
 import java.lang.IllegalArgumentException
 import java.lang.IllegalStateException
 
-class CellTest : BehaviorSpec ({
+class CellTest : BehaviorSpec({
     given("a cell") {
         `when`("is open") {
             then("should not be able to open") {
@@ -17,7 +19,10 @@ class CellTest : BehaviorSpec ({
         `when`("is open") {
             then("should display correct display value") {
                 val adjacentMineCount = 3
-                Cell(state = CellState.Open, adjacentMineCount = adjacentMineCount).displayValue shouldBe adjacentMineCount.toString()
+                Cell(
+                    state = CellState.Open,
+                    adjacentMineCount = adjacentMineCount
+                ).displayValue shouldBe adjacentMineCount.toString()
             }
         }
 
@@ -46,28 +51,3 @@ class CellTest : BehaviorSpec ({
         }
     }
 })
-
-data class Cell(
-    val isMine: Boolean = false,
-    val adjacentMineCount: Int = 0,
-    var state: CellState = CellState.Close
-) {
-    init {
-        require(adjacentMineCount in 0..8) { "adjacent mine count should be in 0..8 $adjacentMineCount of $this" }
-    }
-
-    val displayValue = if (state == CellState.Close) CLOSE_DISPLAY_CHARACTER else adjacentMineCount.toString()
-
-    fun open() {
-        check(state != CellState.Open) { "cannot open a open cell $this" }
-        state = CellState.Open
-    }
-
-    companion object {
-        const val CLOSE_DISPLAY_CHARACTER = "C"
-    }
-}
-
-enum class CellState {
-    Open, Close
-}
